+++ ExpenseTracker/ExpenseTracker/src/ExpenseTracker.Application/CurrencyFolders/Services/CurrencyService.cs
@@ -1,6 +1,9 @@
-﻿using ExpenseTracker.Application.CurrencyFolders.Interface.Infrastructure;
-using ExpenseTracker.Application.CurrencyFolders.Interface.Application;
+﻿using ErrorOr;
 using ExpenseTracker.Application.CurrencyFolders.Data;
+using ExpenseTracker.Application.CurrencyFolders.Interface.Application;
+using ExpenseTracker.Application.CurrencyFolders.Interface.Infrastructure;
+using ExpenseTracker.Application.CurrencyFolders.Services;
+using ExpenseTracker.Application.TransactionFolders.Services;
 using ExpenseTracker.Domain.CurrencyData;
 
 namespace ExpenseTracker.Application.CurrencyFolders.Services
@@ -14,40 +17,87 @@ namespace ExpenseTracker.Application.CurrencyFolders.Services
         {
             _currencyRepository = currencyRepository;
         }
-        public async Task<AllCurrencies> GetCurrenciesAsync(CancellationToken token)
+        public async Task<ErrorOr<AllCurrencies>> GetCurrenciesAsync(CancellationToken token)
         {
             var getCurrencies = await _currencyRepository.GetCurrenciesAsync(token);
 
+            if (getCurrencies.IsError)
+            {
+                return getCurrencies.Errors;
+            }
+
             return new AllCurrencies
             {
-                currencies = getCurrencies
+                currencies = getCurrencies.Value
             };
         }
 
-        public async Task<Currency> GetCurrencyByIdAsync(int currencyId, CancellationToken token)
+        public async Task<ErrorOr<Currency>> GetCurrencyByIdAsync(int currencyId, CancellationToken token)
         {
+            var validation = CurrencyValidator.ValidateCurrencyId(currencyId);
+
+            if (validation.IsError)
+            {
+                return validation.Errors;
+            }
+
             var getCurrency = await _currencyRepository.GetCurrencyByIdAsync(currencyId, token);
+
+            if (getCurrency.IsError) {
+                return getCurrency.Errors;            
+            }
 
             return getCurrency;
         }
 
-        public async Task<Currency> CreateCurrencyAsync(Currency currency, CancellationToken token)
+        public async Task<ErrorOr<Currency>> CreateCurrencyAsync(Currency currency, CancellationToken token)
         {
+            var validation = CurrencyValidator.ValidateForCreate(currency);
+
+            if (validation.IsError)
+            {
+                return validation.Errors;
+            }
+
             var getCurrency = await _currencyRepository.CreateCurrencyAsync(currency, token);
 
             return getCurrency;
         }
 
-        public async Task<bool> UpdateCurrencyAsync(Currency currency, CancellationToken token)
+        public async Task<ErrorOr<Updated>> UpdateCurrencyAsync(Currency currency, CancellationToken token)
         {
+            var validation = CurrencyValidator.ValidateForUpdate(currency);
+
+            if (validation.IsError)
+            {
+                return validation.Errors;
+            }
+
             var status = await _currencyRepository.UpdateCurrencyAsync(currency, token);
+
+            if (status.IsError)
+            {
+                return status.Errors;
+            }
 
             return status;
         }
 
-        public async Task<bool> DeleteCurrencyAsync(int currencyId, CancellationToken token)
+        public async Task<ErrorOr<Deleted>> DeleteCurrencyAsync(int currencyId, CancellationToken token)
         {
+            var validation = CurrencyValidator.ValidateCurrencyId(currencyId);
+
+            if (validation.IsError)
+            {
+                return validation.Errors;
+            }
+
             var status = await _currencyRepository.DeleteCurrencyAsync(currencyId, token);
+
+            if (status.IsError)
+            {
+                return status.Errors;
+            }
 
             return status;
         }
